StoryGenerator/requirements.txt
transformers
python-docx
fpdf
streamlit
pillow